Depend on the aapt2 library rather than doing things ourselves.
This fixes issue #1 by magically doing a better job of parsing manifest
files.

i.e. I could successfully read the versionCode and minSdkVersion from an
obfuscated AAB file to which @silvanfischer gave me access for testing.

While this means that this library is no longer standalone and consumers
now have a transitive dependency on the aapt2 library (which is only
available in the Google Maven repo), that library seems to be relatively
small and straightforward in terms of its dependencies.

I tested this build with the Google Play Android Publisher plugin for
Jenkins and didn't have any weird problems (e.g. crazy Guava issues).